+++ mopidy/__main__.py
@@ -1,7 +1,6 @@
 from __future__ import unicode_literals
 
 import logging
-import optparse
 import os
 import signal
 import sys
@@ -24,11 +23,11 @@ sys.path.insert(
     0, os.path.abspath(os.path.join(os.path.dirname(__file__), '../')))
 
 
-from mopidy import ext
+from mopidy import commands, ext
 from mopidy.audio import Audio
 from mopidy import config as config_lib
 from mopidy.core import Core
-from mopidy.utils import deps, log, path, process, versioning
+from mopidy.utils import log, path, process
 
 logger = logging.getLogger('mopidy.main')
 
@@ -37,11 +36,13 @@ def main():
     signal.signal(signal.SIGTERM, process.exit_handler)
     signal.signal(signal.SIGUSR1, pykka.debug.log_thread_tracebacks)
 
-    loop = gobject.MainLoop()
-    options = parse_options()
-    config_files = options.config.split(b':')
-    config_overrides = options.overrides
+    args = commands.parser.parse_args(args=mopidy_args)
+    if args.show_config:
+        commands.show_config(args)
+    if args.show_deps:
+        commands.show_deps()
 
+    loop = gobject.MainLoop()
     enabled_extensions = []  # Make sure it is defined before the finally block
     logging_initialized = False
 
@@ -50,17 +51,18 @@ def main():
 
     try:
         # Initial config without extensions to bootstrap logging.
-        logging_config, _ = config_lib.load(config_files, [], config_overrides)
+        logging_config, _ = config_lib.load(
+            args.config_files, [], args.config_overrides)
 
         # TODO: setup_logging needs defaults in-case config values are None
         log.setup_logging(
-            logging_config, options.verbosity_level, options.save_debug_log)
+            logging_config, args.verbosity_level, args.save_debug_log)
         logging_initialized = True
 
         installed_extensions = ext.load_extensions()
 
         config, config_errors = config_lib.load(
-            config_files, installed_extensions, config_overrides)
+            args.config_files, installed_extensions, args.config_overrides)
 
         # Filter out disabled extensions and remove any config errors for them.
         for extension in installed_extensions:
@@ -121,78 +123,6 @@ def check_config_errors(errors):
         for key, msg in errors[section].items():
             logger.error('Config value %s/%s %s', section, key, msg)
     sys.exit(1)
-
-
-def check_config_override(option, opt, override):
-    try:
-        return config_lib.parse_override(override)
-    except ValueError:
-        raise optparse.OptionValueError(
-            'option %s: must have the format section/key=value' % opt)
-
-
-def parse_options():
-    parser = optparse.OptionParser(
-        version='Mopidy %s' % versioning.get_version())
-
-    # Ugly extension of optparse type checking magic :/
-    optparse.Option.TYPES += ('config_override',)
-    optparse.Option.TYPE_CHECKER['config_override'] = check_config_override
-
-    parser.add_option(
-        '-q', '--quiet',
-        action='store_const', const=0, dest='verbosity_level',
-        help='less output (warning level)')
-    parser.add_option(
-        '-v', '--verbose',
-        action='count', default=1, dest='verbosity_level',
-        help='more output (debug level)')
-    parser.add_option(
-        '--save-debug-log',
-        action='store_true', dest='save_debug_log',
-        help='save debug log to "./mopidy.log"')
-    parser.add_option(
-        '--show-config',
-        action='callback', callback=show_config_callback,
-        help='show current config')
-    parser.add_option(
-        '--show-deps',
-        action='callback', callback=deps.show_deps_optparse_callback,
-        help='show dependencies and their versions')
-    parser.add_option(
-        '--config',
-        action='store', dest='config',
-        default=b'$XDG_CONFIG_DIR/mopidy/mopidy.conf',
-        help='config files to use, colon seperated, later files override')
-    parser.add_option(
-        '-o', '--option',
-        action='append', dest='overrides', type='config_override',
-        help='`section/key=value` values to override config options')
-    return parser.parse_args(args=mopidy_args)[0]
-
-
-def show_config_callback(option, opt, value, parser):
-    # TODO: don't use callback for this as --config or -o set after
-    # --show-config will be ignored.
-    files = getattr(parser.values, 'config', b'').split(b':')
-    overrides = getattr(parser.values, 'overrides', [])
-
-    extensions = ext.load_extensions()
-    config, errors = config_lib.load(files, extensions, overrides)
-
-    # Clear out any config for disabled extensions.
-    for extension in extensions:
-        if not ext.validate_extension(extension):
-            config[extension.ext_name] = {b'enabled': False}
-            errors[extension.ext_name] = {
-                b'enabled': b'extension disabled its self.'}
-        elif not config[extension.ext_name]['enabled']:
-            config[extension.ext_name] = {b'enabled': False}
-            errors[extension.ext_name] = {
-                b'enabled': b'extension disabled by config.'}
-
-    print config_lib.format(config, extensions, errors)
-    sys.exit(0)
 
 
 def check_old_locations():